Added an EntryPoint, Application and Logging System
Added an EntryPoint, which other applications can use as an interface to the Engine. Added a application class that starts the application loop. Added a cross-platform logging system for debug purposes. Changed some Build.lua files to include the Vendor forlder for included libraries. Changed the MacOS script to setup a project for xcode and added a Git-Clean script to clean the repository from all unneeded files.