Changed Sprite Component and other small things
Changed Sprite Component to change the width and height of a Sprite to match the image or rectangle the Sprite contains. Added a SpriteRenderer System that Culls Sprites that are out of view. Added a new Entity System for Player movement. Added a Tiled Level and Spritesheets to test new Features. Added a LevelLoader that parses the contents of the Tiled levels. Did some small Bug Fixing in various Classes.